GitFetcher: Fix race in detecting changes by using a local branch
Scopes share local git clones (if they're using the same repo).
Change detection was done by comparing the commit hash of the remote branch before and after the fetch.
But repo could've been already fetched for another scope.
To solve that - each remote branch is tracked by a local ref - this local ref is used to determine the old revision.